# Reduce number of `stat` calls in storage `File`