### Changelog category (leave one):
- Performance Improvement

### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
Improve performance of reading from storage `File` and table functions `file` in case when path has globs and matched directory contains large number of files.