@alexey-milovidov 2022-04-25
Thank you!

Further directions:
- the function for glob matching is strange; it will be better to generate files manually instead of converting globs to regexp;
- maybe then we can easily implement https://github.com/ClickHouse/ClickHouse/issues/36316