@alexey-milovidov 2022-04-30
> Stress test (debug, actions) — Found signal in gdb.log 

See this: #36660